app/requirements.txt
fastapi
ipdb
pillow
uvicorn
pydantic
python-multipart
tensorflow
numpy